Create READ FIRST.md
I plan to update the file to have better tracking so that files don't get put into recursive folder structure if organized in the same folder. This is a high priority but I am super busy this week so will try to get this update implemented ASAP this week, but no promises.